+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/EntityApplicationPanelTest.java
@@ -73,7 +73,7 @@ public class EntityApplicationPanelTest {
 		EntityConnectionProvider.CLIENT_CONNECTION_TYPE.set(EntityConnectionProvider.CONNECTION_TYPE_LOCAL);
 		EntityApplicationPanel.builder(TestApplicationModel.class, TestApplicationPanel.class)
 						.user(UNIT_TEST_USER)
-						.domainType(TestDomain.DOMAIN)
+						.domain(TestDomain.DOMAIN)
 						.setUncaughtExceptionHandler(false)
 						.saveDefaultUsername(false)
 						.displayFrame(false)